Add adlfs preprocessing job
The job shows how to use adlfs for retrieving arrow files and storing
them as parquet files all in Azure Blob storage. Unlike the blobfuse
version, this requires the raw arrow files to already exist in the
container.

To run it, just update the .env.example to use its ray job config:
"k8s/rayjob-dataprep-adlfs.yaml" and also add a storage account key
for authentication in the job. From there, just run the
`run-dataprep.ps` script as usual.

Because the .arrow files are large (+200MB), it is not uncommon for
nodes to run out of memory. In the .env.example file, max concurrency
can be set for each stage of the data preprocessing pipeline to better
help constrain the amount of memory used. Right now the max concurrency
is set to 2 to have better resiliency.